Added more methods to the public interface.
Added lock and unlock to the VXI11 public interface.
Added local and remote to the VXI11 public interface.
Added defines for the different flags for the VXI11 public interface.
Added set and get timeout for the pmap methods, this way the developer
can change the timeout to wait for pmap to resolve.